scalable_airflow/requirements.txt
apache-airflow==1.10.9